change function signature

+++ factorial/factorial.go
@@ -1,6 +1,6 @@
 package factorial
 
-func Factorial(x uint) uint {
+func Factorial(x uint8) uint8 {
 	if x == 1 {
 		return 1
 	}
